requirement.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap4==2024.10
Django==5.1.2
django-crispy-forms==2.3
django-pagination==1.0.7
idna==3.10
python-dateutil==2.9.0.post0
pytz==2024.2
relative-datetime==0.1.4
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
